program now asks for save and load destinations

+++ directory.rb
@@ -1,15 +1,22 @@
 @students = []
 @months = ["January", "February", "March", "April", "May", "June", "July", "August", "September", "October", "November", "December"]
-def load_students(filename = "students.csv")
-  file = File.open(filename, "r")
-  file.readlines.each do |line|
-    name, hobbies, height, birthplace, cohort = line.chomp.split(',')
-    @students << {name: name, hobbies: hobbies, height: height, birthplace: birthplace, cohort: cohort}
-    ## the above happens in order of input line, with comma-separation,
-    #so variables that take more than one value can't have commas in them... how to fix this?
+
+def load_students
+  puts "Please enter the name of the file you would like to load"
+  filename = STDIN.gets.chomp
+  if File.exist?(filename)
+    file = File.open(filename, "r")
+    file.readlines.each do |line|
+      name, hobbies, height, birthplace, cohort = line.chomp.split(',')
+      @students << {name: name, hobbies: hobbies, height: height, birthplace: birthplace, cohort: cohort}
+      ## the above happens in order of input line, with comma-separation,
+      #so variables that take more than one value can't have commas in them... how to fix this?
+      puts "students successfully loaded from #{filename}"
+      file.close
+    end
+  elsif !File.exist?(filename)
+    puts "Sorry, #{filename} doesn't exist. Please try again."
   end
-  puts "students successfully loaded from #{filename}"
-  file.close
 end
 
 def input_students
@@ -71,9 +78,9 @@ end
 def print_menu
   puts "1. Input the students"
   puts "2. Show the students"
-  puts "3. Save the list to students.csv"
+  puts "3. Save the list"
   #need to ask user for file to save to
-  puts "4. Load the list from students.csv"
+  puts "4. Load the list"
   #need to ask user for file to load from
   puts "9. Exit"
 end
@@ -107,7 +114,9 @@ def save_students
   # open the file for writing - If we want to write to a file in Ruby,
   # we need to 'open' the file first. The open() method returns us a
   # reference to the file that we can save as a variable called 'file'
-  file = File.open(save_destination = "students.csv", "w")
+  puts "choose the name of your save file"
+  save_destination = STDIN.gets.chomp
+  file = File.open(save_destination, "w")
   #iterate over the array of (hashes of) students
   @students.each do |student|
     # On every iteration we create a new array with the student name and the cohort,
